classifier/requirements.txt
PyGithub
textblob